==> ANN/1. Pre-processing.ipynb <==
# Label each row with features required to run ANN notebooks

The new features are:
* The time left until next bus stop (seconds)
* The time it takes to travel the full segment (seconds)
* The time from the start of the journey to the start of the current segment (seconds)

import numpy as np
import pandas as pds
import datetime as dt
import time
from sklearn import preprocessing as pp

df = pds.read_pickle('train_203_compressed.pkl')
df.head()

#df = pds.read_csv('bus203_all.csv')
#df.head()

Ignore entries that are not `ObservedPositionEvent`

df = df[df['event'] == "ObservedPositionEvent"]

This leaves the indexes of rows untouched, reset.

df = df.reset_index().drop(columns=['index'])

Normalize speed and coordinates

mm = pp.MinMaxScaler()

df['latitude'] = mm.fit_transform(df.latitude.values.reshape(-1,1))
df['longitude'] = mm.fit_transform(df.longitude.values.reshape(-1,1))
df['speed'] = mm.fit_transform(df.speed.values.reshape(-1,1))

Convert `timestamp` to pandas datetime object.

df['timestamp'] = pds.to_datetime(df['timestamp']).dt.tz_localize("UTC").dt.tz_convert("Europe/Stockholm")

A triple loop as it seems, but it is only to finally group rows from each individual segment from every journey. Should be linear in time, as the innermost loop will do all executions and every row is visited once. This only took ~50% of my 8GB of RAM but took ~30 minutes to run.

time_left = pds.DataFrame(np.zeros(len(df.index)), columns=['time_left'])
segment_time = pds.DataFrame(np.zeros(len(df.index)), columns=['segment_time'])
# Time since journey start
tsjs = pds.DataFrame(np.zeros(len(df.index)), columns=['tsjs'])


t0 = time.time()

for j, df_j in df.groupby('journey_number'):
    journey_start = df_j['timestamp'].iloc[0]
    for k, df_s in df_j.groupby('segment_number'):
        end_time = df_s['timestamp'].iloc[-1]
        start_time = df_s['timestamp'].iloc[0]
        for idx, row in df_s.iterrows():
            # The subtraction returns timedelta between the two timestamp objects
            # and total seconds convert the pandas datetime object to seconds
            time_left.iloc[idx] = (end_time - row['timestamp']).total_seconds()
            segment_time.iloc[idx] = (end_time - start_time).total_seconds()
            tsjs.iloc[idx] = (start_time - journey_start).total_seconds()

elapsed = time.time() - t0
print("Data processed in", elapsed, " seconds")

Add new features to dataframe `df`

data = pds.concat([df, time_left,segment_time, tsjs], axis=1)
data.head()

Renaming some columns to keep them similar to the GP model for easier understanding.

data.rename(columns={'longitude': 'lon', 'latitude': 'lat', 'segment_number': 'seg', 'journey_number': 'journey'}, inplace=True)
data.head()

Add history entries for model 3

history_speed = pds.DataFrame(0, index=np.arange(len(data)), columns=['speed_one', 'speed_two', 'speed_three'])
history_pos = pds.DataFrame(0, index=np.arange(len(data)), columns=['lat_one', 'lon_one', 'lat_two', 'lon_two', 'lat_three', 'lon_three'])

Add columns for the three latest positions for each sample. Replace the first sample from each journey with itself.

for i, a in data.groupby('journey'):
    idxs = a.index

    history_pos.lat_one.iloc[idxs] = a.lat.shift(1)
    history_pos.lat_one.iloc[idxs[0]] = history_pos.lat_one.iloc[1]

    history_pos.lon_one.iloc[idxs] = a.lon.shift(1)
    history_pos.lon_one.iloc[idxs[0]] = history_pos.lon_one.iloc[1]

    history_pos.lat_two.iloc[idxs] = a.lat.shift(2)
    history_pos.lat_two.iloc[idxs[0]] = history_pos.lat_two.iloc[2]
    history_pos.lat_two.iloc[idxs[1]] = history_pos.lat_two.iloc[2]

    history_pos.lon_two.iloc[idxs] = a.lon.shift(2)
    history_pos.lon_two.iloc[idxs[0]] = history_pos.lon_two.iloc[2]
    history_pos.lon_two.iloc[idxs[1]] = history_pos.lon_two.iloc[2]

    history_pos.lat_three.iloc[idxs] = a.lat.shift(3)
    history_pos.lat_three.iloc[idxs[0]] = history_pos.lat_three.iloc[3]
    history_pos.lat_three.iloc[idxs[1]] = history_pos.lat_three.iloc[3]
    history_pos.lat_three.iloc[idxs[2]] = history_pos.lat_three.iloc[3]

    history_pos.lon_three.iloc[idxs] = a.lon.shift(3)
    history_pos.lon_three.iloc[idxs[0]] = history_pos.lon_three.iloc[3]
    history_pos.lon_three.iloc[idxs[1]] = history_pos.lon_three.iloc[3]
    history_pos.lon_three.iloc[idxs[2]] = history_pos.lon_three.iloc[3]

and speed

for i, a in data.groupby('journey'):
    idxs = a.index

    history_speed.speed_one.iloc[idxs] = a.lat.shift(1)
    history_speed.speed_one.iloc[idxs[0]] = history_speed.speed_one.iloc[1]

    history_speed.speed_two.iloc[idxs] = a.lat.shift(2)
    history_speed.speed_two.iloc[idxs[0]] = history_speed.speed_two.iloc[2]
    history_speed.speed_two.iloc[idxs[1]] = history_speed.speed_two.iloc[2]

    history_speed.speed_three.iloc[idxs] = a.lat.shift(3)
    history_speed.speed_three.iloc[idxs[0]] = history_speed.speed_three.iloc[3]
    history_speed.speed_three.iloc[idxs[1]] = history_speed.speed_three.iloc[3]
    history_speed.speed_three.iloc[idxs[2]] = history_speed.speed_three.iloc[3]

Make sure there are no NaN:s produced by the shifting

history_pos.isna().sum()

history_speed.isna().sum()

The two cells above should contain only zeros, otherwise, model training will fail!

new_data = pds.concat([history_pos, history_speed], axis=1)
data = pds.concat([data, new_data], axis=1)
data.head()

... remove unwanted columns ...

data = data.drop(columns=['compressed', 'event'])
#data = data.drop(columns=['Unnamed: 0', 'event', 'vehicle_id', 'line', 'station'])
data.head()

... and save it.

data.to_pickle('ANN_train_203_downsampled.pkl')
