==> ANN/2. Extra - parameter optimization.ipynb <==
Note, still in the process of experimenting with parameters

Import packages

import tensorflow as tf
import numpy as np
import pandas as pds
from tensorflow import keras
from hyperas import optim
from hyperas.distributions import choice, uniform
from hyperopt import Trials, STATUS_OK, tpe
from tensorflow.keras.optimizers import Adadelta, Nadam
from tensorflow.keras import Sequential
from tensorflow.keras.layers import Dense

The model is added with wanted combinations to be tested

def create_model(x_train, y_train, x_test, y_test):

    model = Sequential()
    
    model.add(Dense({{choice([3,19,38])}}, activation={{choice([None, 'relu', 'tanh'])}}, input_shape=(19,)))
    model.add(Dense({{choice([1,19,38])}}, activation={{choice([None, 'relu', 'tanh'])}}))
    model.add(Dense({{choice([1,19,38])}}, activation={{choice([None, 'relu', 'tanh'])}}))
    model.add(Dense(1, activation={{choice([None, 'relu'])}}))

    model.compile(loss='mae', 
                  optimizer='Adam',
                  metrics=['mae']
                 )
    
    #result = model.fit(x_train,
    #                  y_train,
    ##                  epochs = 5,
    #                  batch_size=10,
    #                  validation_data=[x_test, y_test])
    result = model.fit(x_train, y_train,
                   batch_size={{choice([64, 128])}},
                   epochs=3,
                   verbose=1,
                   validation_split=0.1)
    
    loss_metric = np.amax(result.history['mean_absolute_error']) 
    print('Current epoch:', result.history['mean_absolute_error'])
    return {'loss': loss_metric, 'status': STATUS_OK, 'model': model}

The model needs a function that deliver the data once in its requested format

def data():
    df = pds.read_pickle('ANN_dataset.pkl')
    
    labels = df['time_left']
    #dataset = dataset.drop(columns=['time_left', 'segment_time']) 
    
    # Make segments categorical
    dataset = pds.get_dummies(df['seg'])

    # Combine segments, timestamps and time from journey start

    # Cyclical timestamps 
    dataset['hr_sin'] = np.sin(df.timestamp.dt.hour*(2.*np.pi/24))
    dataset['hr_cos'] = np.cos(df.timestamp.dt.hour*(2.*np.pi/24))

    # Convert to radians befor trigonometric functions
    dataset['dir_sin'] = np.sin(df.direction*(np.pi/180))
    dataset['dir_cos'] = np.cos(df.direction*(np.pi/180))

    dataset['speed'] = (df['speed']-df['speed'].min())/(df['speed'].max()-df['speed'].min())
    dataset['tsjs'] = (df['tsjs']-df['tsjs'].min())/(df['tsjs'].max()-df['tsjs'].min())

    dataset['lat'] = (df['lat']-df['lat'].min())/(df['lat'].max()-df['lat'].min())
    dataset['lon'] = (df['lon']-df['lon'].min())/(df['lon'].max()-df['lon'].min())
    
    # Include journey number to select entire journeys
    dataset['journey'] = df['journey']
    
    
    divider = 0.2
    num_journeys = dataset.journey.unique()[-1]
    num_test_segments = np.int_(np.round(num_journeys*divider))
    msk = np.random.randint(1, num_journeys, num_test_segments)
    mask = dataset['journey'].isin(msk)

    train_input = dataset[~mask]
    test_input = dataset[mask]

    train_labels = labels[~mask]
    test_labels = labels[mask]

    train_data = train_input.drop(columns=['journey'])
    test_data = test_input.drop(columns=['journey'])

    x_train = train_input.drop(columns=['journey'])
    x_test = test_data
    y_train = train_labels
    y_test = test_labels
    
    return x_train, y_train, x_test, y_test

This calls test, by random choice, 35 different combinations of parameters to the model. The number of evaluations could be increased with the cost of it taking a lot more time.

best_run, best_model = optim.minimize(model=create_model,
                                      data=data,
                                      algo=tpe.suggest,
                                      max_evals=35,
                                      trials=Trials(),
                                      notebook_name="2.\ Extra\ -\ parameter\ optimization"
                                     )

best_run
