==> ANN/2. Model creation - M3.ipynb <==
# Model creation and data split-up

This notebook will:
* Train the model
* Save the model and predictions

Import packages

import tensorflow as tf
from tensorflow import keras
import numpy as np
import pandas as pds
from sklearn import preprocessing as pp

Import data

df_tr = pds.read_pickle('ANN_train_203_downsampled.pkl')
df_te = pds.read_pickle('ANN_test_203_downsampled.pkl')
df = pds.concat([df_tr, df_te], keys=['train', 'test'])
df.head()

#df = pds.read_pickle('ANN_dataset.pkl')
#df.head()

Note that `time_left` is the label in this model 

label_string = 'time_left'

One hot encode segments, normalize columns and create final dataset

# Make segments categorical
dataset = pds.get_dummies(df['seg'])

# Combine segments, timestamps and time from journey start
mm = pp.MinMaxScaler()

# Cyclical timestamps 
dataset['hr_sin'] = np.sin(df.timestamp.dt.hour*(2.*np.pi/24))
dataset['hr_cos'] = np.cos(df.timestamp.dt.hour*(2.*np.pi/24))

# Convert to radians befor trigonometric functions
dataset['dir_sin'] = np.sin(df.direction*(np.pi/180))
dataset['dir_cos'] = np.cos(df.direction*(np.pi/180))


dataset['tsjs'] = (df['tsjs']-df['tsjs'].min())/(df['tsjs'].max()-df['tsjs'].min())

dataset['lat'] = df['lat']
dataset['lon'] = df['lon']
dataset['speed'] = df['speed']

dataset['lat_one'] = df['lat_one']
dataset['lon_one'] = df['lon_one']

dataset['lat_two'] = df['lat_two']
dataset['lon_two'] = df['lon_two']

dataset['lat_three'] = df['lat_three']
dataset['lon_three'] = df['lon_three']

dataset['speed_one'] = df['speed_one']
dataset['speed_two'] = df['speed_two']
dataset['speed_three'] = df['speed_three']

# Include journey number to select entire journeys
dataset['journey'] = df['journey']

dataset.head()

Select an arbitrary number of test journeys

### If the data is not split into train and test, this chunk chould be run
divider = 0.2
num_journeys = dataset.journey.unique()[-1]
num_test_segments = np.int_(np.round(num_journeys*divider))
msk = np.random.randint(1, num_journeys, num_test_segments)
mask = dataset['journey'].isin(msk)

Divide training and test data, where data is a sequence from a journey

### If the data is not split into train and test, this chunk chould be run
train_input = dataset[~mask]
test_input = dataset[mask]

train_labels = df[label_string][~mask]
test_labels = df[label_string][mask]

test_idxs = test_input.index

train_input = dataset.loc['train']
test_input = dataset.loc['test']

train_labels = df.loc['train'][label_string]
test_labels = df.loc['test'][label_string]

test_idxs = test_labels.index

# Journey numbers were needed up until this point
train_data = train_input.drop(columns=['journey'])
test_data = test_input.drop(columns=['journey'])

Build model

model = keras.Sequential([
	keras.layers.Dense(2*len(train_data.columns), activation=tf.nn.relu, input_shape=(train_data.shape[1],)),
    keras.layers.Dense(1*len(train_data.columns), activation=tf.nn.tanh),
    keras.layers.Dense(1*len(train_data.columns)),
	keras.layers.Dense(1, activation=tf.nn.relu)
	])

optimizer = keras.optimizers.Adadelta()
model.compile(loss='mae', optimizer = optimizer, metrics=['mae'])
print(model.summary())

Fit model, 10% of the training data is used as validation set

EPOCHS=10
result = model.fit(train_data.values, 
                   train_labels.values, 
                   epochs = EPOCHS, 
                   validation_split=0.1, 
                   initial_epoch=0,
                   batch_size=32
                  )

Evaluate model on test data

model.evaluate(test_data.values, test_labels.values)

Prepare data for evaluation step

predictions = model.predict(test_data.values)

### Create dataset for evaluation

segment = df['seg'].loc['test'].reset_index(drop=True)

output = pds.DataFrame()
output['seg'] = segment
output['journey'] = test_input['journey'].values
output['speed'] = test_data.speed.values
output['pred'] = predictions
output['label'] = test_labels.values
output.head()

# Same name for all files
model_name = 'ANN_M3_203_downsampled_benchmark'

# serialize model to YAML
model_yaml = model.to_yaml()
with open(model_name + '.yaml', 'w') as yaml_file:
    yaml_file.write(model_yaml)

# Serialize weights to HDF5
model.save_weights(model_name + '.h5')

output.to_pickle(model_name + '.pkl')

Done! Get ready for evaluation
