==> ANN/2. Model creation - M4.ipynb <==
# Model creation and data split-up

This notebook will:
* Train the model
* Save the model and predictions

Import packages

import tensorflow as tf
from tensorflow import keras
import numpy as np
import matplotlib.pyplot as plt
import pandas as pds
import datetime as dt
from collections import deque
import random

Import data

df_tr = pds.read_pickle('ANN_train_203_downsampled.pkl')
df_te = pds.read_pickle('ANN_test_203_downsampled.pkl')
df = pds.concat([df_tr, df_te], keys=['train', 'test'])
df.head()

One hot encode segments, normalize columns. time_left is the label in this model.

dataset = pds.DataFrame()
# Make segments categorical
dataset = pds.get_dummies(df['seg'])

# Combine segments, timestamps and time from journey start

# Cyclical timestamps 

dataset['hr_sin'] = np.sin((pds.to_datetime(df.timestamp).dt.hour*3600+pds.to_datetime(df.timestamp).dt.minute*60+pds.to_datetime(df.timestamp).dt.second)*(2.*np.pi/86400))
dataset['hr_cos'] = np.cos((pds.to_datetime(df.timestamp).dt.hour*3600+pds.to_datetime(df.timestamp).dt.minute*60+pds.to_datetime(df.timestamp).dt.second)*(2.*np.pi/86400))

# Convert to radians befor trigonometric functions
dataset['dir_sin'] = np.sin(df.direction*(np.pi/180))
dataset['dir_cos'] = np.cos(df.direction*(np.pi/180))

dataset['speed'] = df['speed']
# Previously was 'time_since_journey_start'
dataset['tsjs'] = (df['tsjs']-df['tsjs'].min())/(df['tsjs'].max()-df['tsjs'].min())

dataset['lat'] = df['lat']
dataset['lon'] = df['lon']

# Include journey number to select entire journeys
dataset['journey'] = df['journey']

dataset['time_left'] = df['time_left']



dataset.head()

Divide training/validation/test data set

### If the data is not split into train and test, this chunk chould be run
Roughly 20% Test and 20% Validation
num_test_segments = 700
num_validation_segments = 700
unique_journeys = dataset.journey.unique()
random.shuffle(unique_journeys)

msk1 = unique_journeys[0:num_test_segments]
msk2 = unique_journeys[num_test_segments:num_test_segments+num_validation_segments]
msk3 = unique_journeys[num_test_segments+num_validation_segments:len(unique_journeys)]
mask1 = dataset['journey'].isin(msk1)
mask2 = dataset['journey'].isin(msk2)
mask3 = dataset['journey'].isin(msk3)

test_input = dataset.loc['test']
#validation_input = dataset[mask2]
train_input = dataset.loc['train']


# Journey numbers were needed up until this point
train_data = train_input.drop(columns=['journey'])
test_data = test_input.drop(columns=['journey'])
#validation_data = validation_input.drop(columns=['journey'])

train_data.head()

Modify dataset to contain sequences. This is required for the RNN. Each sequence is SEQUENCE_LENGTH consecutive data points. Currently a sequence can contain data points from two different journeys, this is not optimal. Should check if the sequence starts and ends in the same journey. This will also use a lot of RAM since the data is effectively increased by SEQUENCE_LENGTH times.

SEQUENCE_LENGTH = 20
def sequencify(df, shuffle):
    sequencial_data = []
    sequence = deque(maxlen=SEQUENCE_LENGTH)
    for i in df.values:
        sequence.append([n for n in i[:-1]])
        if len(sequence) == SEQUENCE_LENGTH:
            sequencial_data.append([np.array(sequence), i[-1]])
    if shuffle:
        random.shuffle(sequencial_data)
    X = []
    Y = []
    for sequence, label in sequencial_data:
        X.append(sequence)
        Y.append(label)
    return np.array(X),Y

Creating the sequences of training data, test data and validation data. Test data is not randomized in order to be able to plot it in a nice way.

seq_train_x, seq_train_y = sequencify(train_data, True)
seq_test_x, seq_test_y = sequencify(test_data, False)
#seq_validation_x, seq_validation_y = sequencify(validation_data, True)

Model creation

BATCH_SIZE = 64
EPOCHS = 8
model = keras.Sequential()
model.add(keras.layers.CuDNNLSTM(128, input_shape=(seq_train_x.shape[1:]), return_sequences=True))

model.add(keras.layers.CuDNNLSTM(128, input_shape=(seq_train_x.shape[1:]), return_sequences=True))

model.add(keras.layers.CuDNNLSTM(128, input_shape=(seq_train_x.shape[1:])))

model.add(keras.layers.Dense(32, activation="relu"))

model.add(keras.layers.Dense(1, activation="relu"))

opt = keras.optimizers.Adadelta()

model.compile(loss='mae',optimizer=opt, metrics=['mae'])

Fit model

result = model.fit(seq_train_x,seq_train_y,
                  batch_size=BATCH_SIZE,
                  epochs=EPOCHS,
                  validation_split=0.1, )

Evaluate model on all test data

 model.evaluate(seq_test_x, seq_test_y)

Prepare data for the evaluation step

predictions = model.predict(seq_test_x)

### Create dataset for evaluation

segment = df['seg'].loc['test'].reset_index(drop=True)

output = pds.DataFrame()
output['seg'] = segment[19:]
output['journey'] = test_input['journey'][19:].values
output['speed'] = test_data['speed'][19:].values
output['pred'] = predictions
output['label'] = test_data['time_left'][19:].values
output = output.reset_index(drop = True)
output.head()

# Same name for all files
model_name = 'ANN_M4_203_downsampled_benchmark'

# Serialize model to YAML
model_yaml = model.to_yaml()
with open(model_name + '.yaml', "w") as yaml_file:
    yaml_file.write(model_yaml)

# Serialize weights to HDF5
model.save_weights(model_name + '.h5')

output.to_pickle(model_name + '.pkl')

Done! Get ready for evaluation
