==> ANN/4. Model comparison.ipynb <==
# Model comparison
This notebook evaluates a model according to an agreed upon metric in order to compare results with the GP models.

import numpy as np
import pandas as pds
import matplotlib.pyplot as plt

"""
Some loss functions
"""
def basic_mae(pred, label):
    e = abs(pred - label)
    return [e]

def mape(pred, label):

    e =abs((label - pred)/label)

    return [e]

Pick the name for your model

model_name = 'ANN_M1_203_benchmark'

data = pds.read_pickle(model_name + '.pkl')
data.head()

jour_frames = []

for ij, jour in data.groupby('journey'):
    seg_frames = []
    
    for iss, seg in jour.groupby('seg'):
        tot = len(seg) - 1
        errs = {}
        i_20 = np.int_(np.round(tot * 0.2))
        i_40 = np.int_(np.round(tot * 0.4))
        i_60 = np.int_(np.round(tot * 0.6))
        i_80 = np.int_(np.round(tot * 0.8))
        
        p_20 = seg.iloc[i_20]
        p_40 = seg.iloc[i_40]
        p_60 = seg.iloc[i_60]
        p_80 = seg.iloc[i_80]
        errs['journey'] = [ij]
        errs['segment'] = [iss]
        
        errs['e_20'] = basic_mae(p_20.pred, p_20.label)
        errs['ep_20'] = mape(p_20.pred, p_20.label)
        
        errs['e_40'] = basic_mae(p_40.pred, p_40.label)
        errs['ep_40'] = mape(p_40.pred, p_40.label)
        
        errs['e_60'] = basic_mae(p_60.pred, p_60.label)
        errs['ep_60'] = mape(p_60.pred, p_60.label)
        
        errs['e_80'] = basic_mae(p_80.pred, p_80.label)
        errs['ep_80'] = mape(p_80.pred, p_80.label)
        
        seg_frames.append(pds.DataFrame(errs))
    
    jour_frames.append(pds.concat(seg_frames))
        

ers = pds.concat(jour_frames).reset_index(drop=True)

#ers = pds.read_pickle('ANN_M2_downsampled_203_errors.pkl')

ers.head(30)

m = ers.mean(axis=0)
for i in range(2,9,2):
    print('&',str(round(m[i], 2)), end='')

for i in range(3,10,2):
    print('&',str(round(m[i]*100, 2)), end='')

mae = []
mape = []
for i in ers.segment.unique():
    seg_data = ers.loc[ers['segment'] == i]
    e_20 = seg_data['e_20']
    e_40 = seg_data['e_40']
    e_60 = seg_data['e_60']
    e_80 = seg_data['e_80']
    ep_20 = seg_data['ep_20']
    ep_40 = seg_data['ep_40']
    ep_60 = seg_data['ep_60']
    ep_80 = seg_data['ep_80']
    mae_i = e_20.sum()+e_40.sum()+e_60.sum()+e_80.sum()
    mae_i = mae_i/(len(seg_data)*4)
    mape_i = ep_20.sum()+ep_40.sum()+ep_60.sum()+ep_80.sum()
    mape_i = mape_i/(len(seg_data)*4)
    mae.append(mae_i)
    mape.append(mape_i)

#LaTeX formatting
for i in mae:
    print('&',str(round(i, 2)), end='')
print('\n')
for i in mape:
    print('&',str(round(i*100, 2)), end='')

ers.to_pickle(model_name+'_errors.pkl')
